refactoring and tests update

+++ hometask05/src/main/java/ru/otus/spring/dao/ext/AuthorBooksResultSetExtractor.java
@@ -14,20 +14,20 @@ public class AuthorBooksResultSetExtractor implements ResultSetExtractor<List<Bo
 
     @Override
     public List<Book> extractData(ResultSet rs) throws SQLException, DataAccessException {
-        Map<Integer, Book> books = new HashMap<>();
+        Map<Long, Book> books = new HashMap<>();
         while (rs.next()) {
-            int id = rs.getInt("id");
+            long id = rs.getLong("id");
             Book book = books.get(id);
             if (book == null) {
                 String name = rs.getString("name");
-                int genreId = rs.getInt("genre_Id");
+                long genreId = rs.getLong("genre_Id");
                 String genre = rs.getString("genre");
                 List<Genre> genres = new ArrayList<>();
                 genres.add(new Genre(genreId, genre));
                 book = new Book(id, name, null, genres);
                 books.put(id, book);
             } else {
-                int genreId = rs.getInt("genre_Id");
+                long genreId = rs.getLong("genre_Id");
                 String genre = rs.getString("genre");
                 book.getGenres().add(new Genre(genreId, genre));
             }
