+++ hometask05/src/main/java/ru/otus/spring/domain/Genre.java
@@ -2,15 +2,15 @@ package ru.otus.spring.domain;
 
 public class Genre {
 
-    private final Integer id;
+    private final Long id;
     private final String name;
 
-    public Genre(Integer id, String name) {
+    public Genre(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
